Loop in description of commit
In if statment language that is generating is checked. If the language
is in condition of if statement, then bad_case is called with one
parameter, othewise it is called with two parameters. Second parameter
is select_icon_number variable, which holds number of current Select
icon.
AutoHotkey code generator updated to beta v1.6. Autohotkey documentation
updated. Examples for AutoHotkey generated with updated code generator.